requirements.txt
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0


numpy~=1.24
pandas~=2.0

opencv-python~=4.8
faiss-cpu~=1.8
easyocr>=1.7.1 

openai-whisper==20231117
triton==2.1.0

open-clip-torch~=2.24
git+https://github.com/openai/CLIP.git

pymongo~=4.13.2
transformers==4.57.1
scenedetect~=0.6
qwen-vl-utils==0.0.14

websocket-client~=1.7
websockets~=12.0
requests~=2.31
python-dotenv~=1.0.1


